refactor(ledger): separate token mods
