+++ rust/src/ledger/token/address.rs
@@ -1,0 +1,134 @@
+//! Token address
+
+use super::TokenId;
+use crate::{
+    constants::MINA_TOKEN_ADDRESS, protocol::serialization_types::version_bytes::TOKEN_ID_KEY,
+    utility::store::common::U64_LEN,
+};
+use anyhow::bail;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize)]
+pub struct TokenAddress(pub String);
+
+impl TokenAddress {
+    pub const LEN: usize = 50;
+
+    pub fn from_bytes(bytes: Vec<u8>) -> anyhow::Result<Self> {
+        String::from_utf8(bytes)
+            .map(Self)
+            .map_err(anyhow::Error::new)
+    }
+
+    pub fn new<T>(token: T) -> Option<Self>
+    where
+        T: Into<String>,
+    {
+        let token: String = token.into();
+        if token.len() == Self::LEN {
+            return Some(Self(token));
+        }
+
+        None
+    }
+
+    /// Used to surpass the [TokenAddress] bytes in a db key
+    pub fn upper_bound() -> [u8; TokenAddress::LEN] {
+        [u8::MAX; TokenAddress::LEN]
+    }
+}
+
+///////////
+// serde //
+///////////
+
+impl<'de> Deserialize<'de> for TokenAddress {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        crate::utility::serde::from_str(deserializer)
+    }
+}
+
+/////////////////
+// conversions //
+/////////////////
+
+impl std::str::FromStr for TokenAddress {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if let Some(token) = Self::new(s) {
+            return Ok(token);
+        }
+
+        bail!("Invalid token address: {s}")
+    }
+}
+
+/////////////////
+// conversions //
+/////////////////
+
+impl From<TokenId> for TokenAddress {
+    fn from(value: TokenId) -> Self {
+        let mut big_int = [0; 32];
+        let le_bytes = value.0.to_le_bytes();
+
+        // big int LE bytes
+        big_int[..U64_LEN].copy_from_slice(&le_bytes);
+        Self(
+            bs58::encode(&big_int[..])
+                .with_check_version(TOKEN_ID_KEY)
+                .into_string(),
+        )
+    }
+}
+
+/////////////
+// default //
+/////////////
+
+impl std::default::Default for TokenAddress {
+    /// MINA token address
+    fn default() -> Self {
+        Self(MINA_TOKEN_ADDRESS.into())
+    }
+}
+
+/////////////
+// display //
+/////////////
+
+impl std::fmt::Display for TokenAddress {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::ledger::token::id::TokenId;
+
+    #[test]
+    fn roundtrip() -> anyhow::Result<()> {
+        let token = TokenAddress::default();
+
+        // serialize
+        let ser = serde_json::to_vec(&token)?;
+
+        // deserialize
+        let res: TokenAddress = serde_json::from_slice(&ser)?;
+
+        assert_eq!(token, res);
+        Ok(())
+    }
+
+    #[test]
+    fn id_to_token_address() {
+        let token = TokenAddress::from(TokenId::default());
+        assert_eq!(token.0, MINA_TOKEN_ADDRESS);
+    }
+}
